+++ src/syn_antd/date_picker.cljs
@@ -7,6 +7,8 @@
 
 (def date-picker-month-picker (reagent.core/adapt-react-class (.-MonthPicker ant-date-picker)))
 
+(def date-picker-quarter-picker (reagent.core/adapt-react-class (.-QuarterPicker ant-date-picker)))
+
 (def date-picker-range-picker (reagent.core/adapt-react-class (.-RangePicker ant-date-picker)))
 
 (def date-picker-week-picker (reagent.core/adapt-react-class (.-WeekPicker ant-date-picker)))