+++ src/syn_antd/space.cljs
@@ -1,0 +1,6 @@
+(ns syn-antd.space
+  (:require
+    [reagent.core]
+    ["antd/es/space" :default ant-space]))
+
+(def space (reagent.core/adapt-react-class ant-space))